Fixed an issue with custom car files
If there are two files (one png, one txt) with different names, and they are the only two custom car files, the mod no longer tries to add them as custom cars (i.e. the mod reverts back to the default car). This used to cause an error, as the mod thought there was a valid custom car, but there was not.